goschedstats: refactor code
We now have a single instance of the public APIs.

We also move the definition of the cluster setting so that it exists
in all builds (even if it doesn't do anything, a server built without
our runtime should recognize the setting name).
